+++ src/__tests__/integration/Client-onReady.spec.ts
@@ -32,7 +32,7 @@ describe('onReady Handler', () => {
 			expect(zbc2.connected).toBe(true)
 			await zbc2.close()
 			done()
-		}, 4000)
+		}, 6000)
 	})
 
 	it(`Does emit the ready event if there is a broker`, done => {
@@ -46,6 +46,6 @@ describe('onReady Handler', () => {
 			expect(zbc2.connected).toBe(true)
 			await zbc2.close()
 			done()
-		}, 4000)
+		}, 6000)
 	})
 })
