+++ src/lib/MockStdOut.ts
@@ -1,4 +1,4 @@
-import { ZBCustomLogger } from './interfaces'
+import { ZBCustomLogger } from './interfaces-published-contract'
 export class MockStdOut implements ZBCustomLogger {
 	public messages: string[] = []
